+++ numpy/src/main/java9/module-info.java
@@ -1,0 +1,5 @@
+module org.bytedeco.numpy {
+  requires transitive org.bytedeco.javacpp;
+  exports org.bytedeco.numpy.global;
+  exports org.bytedeco.numpy;
+}